Move uniffi impl
