+++ nym-vpn-lib/src/uniffi_custom_impls.rs
@@ -5,6 +5,7 @@ use crate::platform::error::FFIError;
 use crate::{NodeIdentity, Recipient, UniffiCustomTypeConverter};
 use ipnetwork::IpNetwork;
 use nym_explorer_client::Location;
+use nym_gateway_directory::{EntryPoint, ExitPoint};
 use std::net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr};
 use std::str::FromStr;
 use talpid_types::net::wireguard::{PresharedKey, PrivateKey, PublicKey};
@@ -22,6 +23,8 @@ uniffi::custom_type!(Url, String);
 uniffi::custom_type!(NodeIdentity, String);
 uniffi::custom_type!(Recipient, String);
 uniffi::custom_type!(Location, String);
+uniffi::custom_type!(EntryPoint, String);
+uniffi::custom_type!(ExitPoint, String);
 
 impl UniffiCustomTypeConverter for NodeIdentity {
     type Builtin = String;
@@ -170,5 +173,29 @@ impl UniffiCustomTypeConverter for Location {
 
     fn from_custom(obj: Self) -> Self::Builtin {
         serde_json::to_string(&obj).unwrap()
+    }
+}
+
+impl UniffiCustomTypeConverter for EntryPoint {
+    type Builtin = String;
+
+    fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
+        Ok(serde_json::from_str(&val)?)
+    }
+
+    fn from_custom(obj: Self) -> Self::Builtin {
+        obj.to_string()
+    }
+}
+
+impl UniffiCustomTypeConverter for ExitPoint {
+    type Builtin = String;
+
+    fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
+        Ok(serde_json::from_str(&val)?)
+    }
+
+    fn from_custom(obj: Self) -> Self::Builtin {
+        obj.to_string()
     }
 }
